# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RISKOS - AI Payment Risk Engine",
    "description": "Autonomous payment fraud detection, AI investigation, entity graph, and counterfactual simulation platform.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "List Transactions",
        "description": "List transactions with pagination and optional filtering.",
        "operationId": "list_transactions_api_transactions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset"
          },
          {
            "name": "customer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Ingest Transaction",
        "description": "Ingest a new payment transaction.",
        "operationId": "ingest_transaction_api_transactions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get Transaction",
        "description": "Retrieve a single transaction by UUID.",
        "operationId": "get_transaction_api_transactions__transaction_id__get",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/investigations": {
      "post": {
        "tags": [
          "Investigations"
        ],
        "summary": "Trigger Investigation",
        "description": "Trigger an AI investigation for a transaction.",
        "operationId": "trigger_investigation_api_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}": {
      "get": {
        "tags": [
          "Investigations"
        ],
        "summary": "Get Investigation",
        "description": "Fetch investigation detail.",
        "operationId": "get_investigation_api_investigations__investigation_id__get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reviews/pending": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "List Pending Reviews",
        "description": "List investigations awaiting human decision with pagination.",
        "operationId": "list_pending_reviews_api_reviews_pending_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingReviewsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{investigation_id}": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Submit Review",
        "description": "Submit analyst decision on an investigation and update transaction status accordingly.",
        "operationId": "submit_review_api_reviews__investigation_id__post",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FinalAction": {
        "type": "string",
        "enum": [
          "allow",
          "verify",
          "hold",
          "block"
        ],
        "title": "FinalAction"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvestigationCreate": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "transaction_id"
        ],
        "title": "InvestigationCreate"
      },
      "InvestigationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "evidence": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "ai_conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Conclusion"
          },
          "recommended_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendedAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "transaction_id",
          "created_at"
        ],
        "title": "InvestigationResponse"
      },
...
      "PendingReviewItem": {
        "properties": {
          "investigation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Investigation Id"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Transaction Id"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "amount_inr": {
            "type": "number",
            "title": "Amount Inr"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "investigation_id",
          "transaction_id",
          "amount",
          "amount_inr",
          "created_at"
        ],
        "title": "PendingReviewItem"
      },
      "PendingReviewsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PendingReviewItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PendingReviewsResponse"
      },
      "RecommendedAction": {
        "type": "string",
        "enum": [
          "allow",
          "verify",
          "hold",
          "block"
        ],
        "title": "RecommendedAction"
      },
      "ReviewCreate": {
        "properties": {
          "reviewer_name": {
            "type": "string",
            "title": "Reviewer Name"
          },
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "final_action": {
            "$ref": "#/components/schemas/FinalAction"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reviewer_name",
          "decision",
          "final_action"
        ],
        "title": "ReviewCreate"
      },
      "ReviewDecision": {
        "type": "string",
        "enum": [
          "approved_ai_recommendation",
          "overridden"
        ],
        "title": "ReviewDecision"
      },
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "investigation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Investigation Id"
          },
          "reviewer_name": {
            "type": "string",
            "title": "Reviewer Name"
          },
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision"
          },
          "final_action": {
            "$ref": "#/components/schemas/FinalAction"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "investigation_id",
          "reviewer_name",
          "decision",
          "final_action",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
      "RiskLabel": {
        "type": "string",
        "enum": [
          "allow",
          "verify",
          "hold",
          "block"
        ],
        "title": "RiskLabel"
      },
...
      "TransactionCreate": {
        "properties": {
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          },
          "merchant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Merchant Id"
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "description": "Transaction amount in paise (must be greater than 0)"
          },
          "currency": {
            "type": "string",
            "maxLength": 10,
            "title": "Currency",
            "default": "INR"
          },
          "device_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Fingerprint"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "instrument_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Hash"
          },
          "instrument_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Type"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "merchant_id",
          "amount"
        ],
        "title": "TransactionCreate"
      },
      "TransactionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "TransactionListResponse"
      },
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          },
          "merchant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Merchant Id"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "$ref": "#/components/schemas/TransactionStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Score"
          },
          "risk_label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskLabel"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount_inr": {
            "type": "number",
            "title": "Amount Inr",
            "description": "Amount in INR (paise / 100)",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "merchant_id",
          "amount",
          "currency",
          "status",
          "created_at"
        ],
        "title": "TransactionResponse"
      },
      "TransactionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "declined",
          "disputed",
          "held"
        ],
        "title": "TransactionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}